+++ crates/swc_css_minifier/src/compress/mod.rs
@@ -11,4 +11,3 @@ pub mod length;
 pub mod selector;
 pub mod time;
 pub mod transform_function;
-pub mod unicore_range;
