+++ crates/swc_css_minifier/src/lib.rs
@@ -11,7 +11,7 @@ use self::{
         easing_function::compress_easing_function, empty::compress_empty,
         frequency::compress_frequency, keyframes::compress_keyframes, length::compress_length,
         selector::compress_selector, time::compress_time,
-        transform_function::compress_transform_function, unicore_range::compress_unicode_range,
+        transform_function::compress_transform_function,
     },
     options::MinifyOptions,
 };
@@ -28,7 +28,6 @@ pub fn minify(stylesheet: &mut Stylesheet, _options: MinifyOptions) {
     stylesheet.visit_mut_with(&mut compress_angle());
     stylesheet.visit_mut_with(&mut compress_time());
     stylesheet.visit_mut_with(&mut compress_frequency());
-    stylesheet.visit_mut_with(&mut compress_unicode_range());
     stylesheet.visit_mut_with(&mut compress_easing_function());
     stylesheet.visit_mut_with(&mut compress_transform_function());
     stylesheet.visit_mut_with(&mut compress_declaration());
